ipam: Factor out IPAM related metrics
These metrics will be reused by additional upcoming IPAM implementations
