+++ pkg/aws/eni/metrics/metrics.go
@@ -17,28 +17,15 @@ package metrics
 import (
 	"time"
 
-	"github.com/cilium/cilium/pkg/trigger"
-
 	"github.com/prometheus/client_golang/prometheus"
 )
 
 const eniSubsystem = "eni"
 
 type prometheusMetrics struct {
-	registry              *prometheus.Registry
-	AllocateEniOps        *prometheus.CounterVec
-	AllocateIpOps         *prometheus.CounterVec
-	ReleaseIpOps          *prometheus.CounterVec
-	IPsAllocated          *prometheus.GaugeVec
-	AvailableENIs         prometheus.Gauge
-	AvailableIPsPerSubnet *prometheus.GaugeVec
-	Nodes                 *prometheus.GaugeVec
-	Resync                prometheus.Counter
-	EC2ApiDuration        *prometheus.HistogramVec
-	EC2RateLimit          *prometheus.HistogramVec
-	poolMaintainer        *triggerMetrics
-	k8sSync               *triggerMetrics
-	resync                *triggerMetrics
+	registry       *prometheus.Registry
+	EC2ApiDuration *prometheus.HistogramVec
+	EC2RateLimit   *prometheus.HistogramVec
 }
 
 // NewPrometheusMetrics returns a new ENI metrics implementation backed by
@@ -48,68 +35,12 @@ func NewPrometheusMetrics(namespace string, registry *prometheus.Registry) *prom
 		registry: registry,
 	}
 
-	m.IPsAllocated = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "ips",
-		Help:      "Number of IPs allocated",
-	}, []string{"type"})
-
-	m.AllocateIpOps = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "allocation_ops",
-		Help:      "Number of IP allocation operations",
-	}, []string{"subnetId"})
-
-	m.ReleaseIpOps = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "release_ops",
-		Help:      "Number of IP release operations",
-	}, []string{"subnetId"})
-
-	m.AllocateEniOps = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "interface_creation_ops",
-		Help:      "Number of ENIs allocated",
-	}, []string{"subnetId", "status"})
-
-	m.AvailableENIs = prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "available",
-		Help:      "Number of ENIs with addresses available",
-	})
-
-	m.AvailableIPsPerSubnet = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "available_ips_per_subnet",
-		Help:      "Number of available IPs per subnet ID",
-	}, []string{"subnetId", "availabilityZone"})
-
-	m.Nodes = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "nodes",
-		Help:      "Number of nodes by category { total | in-deficit | at-capacity }",
-	}, []string{"category"})
-
 	m.EC2ApiDuration = prometheus.NewHistogramVec(prometheus.HistogramOpts{
 		Namespace: namespace,
 		Subsystem: eniSubsystem,
 		Name:      "aws_api_duration_seconds",
 		Help:      "Duration of interactions with AWS API",
 	}, []string{"operation", "responseCode"})
-
-	m.Resync = prometheus.NewCounter(prometheus.CounterOpts{
-		Namespace: namespace,
-		Subsystem: eniSubsystem,
-		Name:      "resync_total",
-		Help:      "Number of resync operations to synchronize AWS EC2 metadata",
-	})
 
 	m.EC2RateLimit = prometheus.NewHistogramVec(prometheus.HistogramOpts{
 		Namespace: namespace,
@@ -118,67 +49,10 @@ func NewPrometheusMetrics(namespace string, registry *prometheus.Registry) *prom
 		Help:      "Duration of EC2 client-side rate limiter blocking",
 	}, []string{"operation"})
 
-	// pool_maintainer is a more generic name, but for backward compatibility
-	// of dashboard, keep the metric name deficit_resolver unchanged
-	m.poolMaintainer = newTriggerMetrics(namespace, "deficit_resolver")
-	m.k8sSync = newTriggerMetrics(namespace, "k8s_sync")
-	m.resync = newTriggerMetrics(namespace, "ec2_resync")
-
-	registry.MustRegister(m.IPsAllocated)
-	registry.MustRegister(m.AllocateIpOps)
-	registry.MustRegister(m.ReleaseIpOps)
-	registry.MustRegister(m.AllocateEniOps)
-	registry.MustRegister(m.AvailableENIs)
-	registry.MustRegister(m.AvailableIPsPerSubnet)
-	registry.MustRegister(m.Nodes)
-	registry.MustRegister(m.Resync)
 	registry.MustRegister(m.EC2ApiDuration)
 	registry.MustRegister(m.EC2RateLimit)
-	m.poolMaintainer.register(registry)
-	m.k8sSync.register(registry)
-	m.resync.register(registry)
 
 	return m
-}
-
-func (p *prometheusMetrics) PoolMaintainerTrigger() trigger.MetricsObserver {
-	return p.poolMaintainer
-}
-
-func (p *prometheusMetrics) K8sSyncTrigger() trigger.MetricsObserver {
-	return p.k8sSync
-}
-
-func (p *prometheusMetrics) ResyncTrigger() trigger.MetricsObserver {
-	return p.resync
-}
-
-func (p *prometheusMetrics) IncENIAllocationAttempt(status, subnetID string) {
-	p.AllocateEniOps.WithLabelValues(subnetID, status).Inc()
-}
-
-func (p *prometheusMetrics) AddIPAllocation(subnetID string, allocated int64) {
-	p.AllocateIpOps.WithLabelValues(subnetID).Add(float64(allocated))
-}
-
-func (p *prometheusMetrics) AddIPRelease(subnetID string, released int64) {
-	p.ReleaseIpOps.WithLabelValues(subnetID).Add(float64(released))
-}
-
-func (p *prometheusMetrics) SetAllocatedIPs(typ string, allocated int) {
-	p.IPsAllocated.WithLabelValues(typ).Set(float64(allocated))
-}
-
-func (p *prometheusMetrics) SetAvailableENIs(available int) {
-	p.AvailableENIs.Set(float64(available))
-}
-
-func (p *prometheusMetrics) SetAvailableIPsPerSubnet(subnetID string, availabilityZone string, available int) {
-	p.AvailableIPsPerSubnet.WithLabelValues(subnetID, availabilityZone).Set(float64(available))
-}
-
-func (p *prometheusMetrics) SetNodes(label string, nodes int) {
-	p.Nodes.WithLabelValues(label).Set(float64(nodes))
 }
 
 func (p *prometheusMetrics) ObserveEC2APICall(operation, status string, duration float64) {
@@ -187,61 +61,4 @@ func (p *prometheusMetrics) ObserveEC2APICall(operation, status string, duration
 
 func (p *prometheusMetrics) ObserveEC2RateLimit(operation string, delay time.Duration) {
 	p.EC2RateLimit.WithLabelValues(operation).Observe(delay.Seconds())
-}
-
-func (p *prometheusMetrics) IncResyncCount() {
-	p.Resync.Inc()
-}
-
-type triggerMetrics struct {
-	total        prometheus.Counter
-	folds        prometheus.Gauge
-	callDuration prometheus.Histogram
-	latency      prometheus.Histogram
-}
-
-func newTriggerMetrics(namespace, name string) *triggerMetrics {
-	return &triggerMetrics{
-		total: prometheus.NewCounter(prometheus.CounterOpts{
-			Namespace: namespace,
-			Subsystem: eniSubsystem,
-			Name:      name + "_queued_total",
-			Help:      "Number of queued triggers",
-		}),
-		folds: prometheus.NewGauge(prometheus.GaugeOpts{
-			Namespace: namespace,
-			Subsystem: eniSubsystem,
-			Name:      name + "_folds",
-			Help:      "Current level of folding",
-		}),
-		callDuration: prometheus.NewHistogram(prometheus.HistogramOpts{
-			Namespace: namespace,
-			Subsystem: eniSubsystem,
-			Name:      name + "_duration_seconds",
-			Help:      "Duration of trigger runs",
-		}),
-		latency: prometheus.NewHistogram(prometheus.HistogramOpts{
-			Namespace: namespace,
-			Subsystem: eniSubsystem,
-			Name:      name + "_latency_seconds",
-			Help:      "Latency between queue and trigger run",
-		}),
-	}
-}
-
-func (t *triggerMetrics) register(registry *prometheus.Registry) {
-	registry.MustRegister(t.total)
-	registry.MustRegister(t.folds)
-	registry.MustRegister(t.callDuration)
-	registry.MustRegister(t.latency)
-}
-
-func (t *triggerMetrics) QueueEvent(reason string) {
-	t.total.Inc()
-}
-
-func (t *triggerMetrics) PostRun(duration, latency time.Duration, folds int) {
-	t.callDuration.Observe(duration.Seconds())
-	t.latency.Observe(latency.Seconds())
-	t.folds.Set(float64(folds))
 }
