+++ pkg/aws/eni/node_manager_test.go
@@ -23,10 +23,10 @@ import (
 	"time"
 
 	ec2mock "github.com/cilium/cilium/pkg/aws/ec2/mock"
-	metricsmock "github.com/cilium/cilium/pkg/aws/eni/metrics/mock"
 	eniTypes "github.com/cilium/cilium/pkg/aws/eni/types"
 	"github.com/cilium/cilium/pkg/aws/types"
 	"github.com/cilium/cilium/pkg/checker"
+	metricsmock "github.com/cilium/cilium/pkg/ipam/metrics/mock"
 	v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/testutils"
@@ -600,12 +600,12 @@ func (e *ENISuite) TestNodeManagerManyNodes(c *check.C) {
 
 	// All subnets must have been used for allocation
 	for _, subnet := range subnets {
-		c.Assert(metricsapi.ENIAllocationAttempts("success", subnet.ID), check.Not(check.Equals), 0)
+		c.Assert(metricsapi.AllocationAttempts("success", subnet.ID), check.Not(check.Equals), 0)
 		c.Assert(metricsapi.IPAllocations(subnet.ID), check.Not(check.Equals), 0)
 	}
 
 	c.Assert(metricsapi.ResyncCount(), check.Not(check.Equals), 0)
-	c.Assert(metricsapi.AvailableENIs(), check.Not(check.Equals), 0)
+	c.Assert(metricsapi.AvailableInterfaces(), check.Not(check.Equals), 0)
 }
 
 // TestNodeManagerInstanceNotRunning verifies that allocation correctly detects
@@ -638,7 +638,7 @@ func (e *ENISuite) TestNodeManagerInstanceNotRunning(c *check.C) {
 	}, 5*time.Second), check.IsNil)
 
 	// Metric should not indicate failure
-	c.Assert(metricsMock.ENIAllocationAttempts("ENI attachment failed", testSubnet.ID), check.Equals, int64(0))
+	c.Assert(metricsMock.AllocationAttempts("ENI attachment failed", testSubnet.ID), check.Equals, int64(0))
 
 	node := mngr.Get("node1")
 	c.Assert(node, check.Not(check.IsNil))
