+++ pkg/ipam/metrics/mock/mock_test.go
@@ -1,4 +1,4 @@
-// Copyright 2019 Authors of Cilium
+// Copyright 2019-2020 Authors of Cilium
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@ package mock
 
 import (
 	"testing"
-	"time"
 
 	"gopkg.in/check.v1"
 )
@@ -33,9 +32,17 @@ var _ = check.Suite(&MockSuite{})
 
 func (e *MockSuite) TestMock(c *check.C) {
 	api := NewMockMetrics()
-	api.ObserveEC2APICall("DescribeNetworkInterfaces", "success", 2.0)
-	c.Assert(api.EC2APICall("DescribeNetworkInterfaces", "success"), check.Equals, 2.0)
-	api.ObserveEC2RateLimit("DescribeNetworkInterfaces", time.Second)
-	api.ObserveEC2RateLimit("DescribeNetworkInterfaces", time.Second)
-	c.Assert(api.EC2RateLimit("DescribeNetworkInterfaces"), check.Equals, 2*time.Second)
+	api.IncAllocationAttempt("foo", "s-1")
+	c.Assert(api.AllocationAttempts("foo", "s-1"), check.Equals, int64(1))
+	api.AddIPAllocation("s-1", 10)
+	api.AddIPAllocation("s-1", 20)
+	c.Assert(api.IPAllocations("s-1"), check.Equals, int64(30))
+	api.SetAllocatedIPs("used", 200)
+	c.Assert(api.AllocatedIPs("used"), check.Equals, 200)
+	api.SetAvailableInterfaces(10)
+	c.Assert(api.AvailableInterfaces(), check.Equals, 10)
+	api.SetNodes("at-capacity", 5)
+	c.Assert(api.Nodes("at-capacity"), check.Equals, 5)
+	api.IncResyncCount()
+	c.Assert(api.ResyncCount(), check.Equals, int64(1))
 }
